+++ test/node_api_test.dart
@@ -69,13 +69,6 @@ a {
 }'''));
     });
 
-    test("data and file may not both be set", () {
-      var error =
-          renderSyncError(new RenderOptions(data: "x {y: z}", file: sassPath));
-      expect(error.toString(),
-          contains('options.data and options.file may not both be set.'));
-    });
-
     test("one of data and file must be set", () {
       var error = renderSyncError(new RenderOptions());
       expect(error.toString(),
@@ -183,6 +176,37 @@ a {
 a {
  b: c;
 }'''));
+      });
+    });
+
+    group("with both data and file", () {
+      test("uses the data parameter as the source", () {
+        expect(renderSync(new RenderOptions(data: "x {y: z}", file: sassPath)),
+            equalsIgnoringWhitespace('x { y: z; }'));
+      });
+
+      test("doesn't require the file path to exist", () {
+        expect(
+            renderSync(new RenderOptions(
+                data: "x {y: z}", file: p.join(sandbox, 'non-existent.scss'))),
+            equalsIgnoringWhitespace('x { y: z; }'));
+      });
+
+      test("imports relative to the file path", () async {
+        await writeTextFile(p.join(sandbox, 'importee.scss'), 'x {y: z}');
+        expect(
+            renderSync(
+                new RenderOptions(data: "@import 'importee'", file: sassPath)),
+            equalsIgnoringWhitespace('x { y: z; }'));
+      });
+
+      test("reports errors from the file path", () {
+        var error =
+            renderSyncError(new RenderOptions(data: "x {y: }", file: sassPath));
+        expect(
+            error.toString(),
+            equals("Error: Expected expression.\n"
+                "  $sassPath 1:7  root stylesheet"));
       });
     });